refactor(frontend): decompose mega widgets in main app tabs
- Decomposed `ProductivityScreen` into `WorkspaceHeader`, `WorkspaceTabs`, `TaskFilters`, `WorkspaceEmptyState`, `TodayPlanCard`, and `EventCard`.
- Decomposed `SettingsScreen` into `SectionHeader`, `SettingRow`, `MemoryItem`, and `LanguagePickerModal`.
- Decomposed `AgentsScreen` into `AgentsHeader` and `AgentsFilters`.
- Resolved > 500-line "God Class" and "Mega Widget" code smells by enforcing modular UI architecture and the Single Responsibility Principle.
- Eliminated inline view building, separating state and layout logic from complex render methods.
- Verified test suite and TypeScript types remain 100% green with strict no-behavior-change constraints.
