+++ frontend/src/components/chat/ChatListHeader.tsx
@@ -118,9 +118,7 @@ export function ChatListHeader({
                 key={mode}
                 onPress={() => onChangeSortMode(mode)}
                 className={`me-2 rounded-full px-3 py-2 border ${
-                  selected
-                    ? 'bg-[#DDF4EB] border-[#147D6473]'
-                    : 'bg-[#ECF5F0] border-[#DDE8E0]'
+                  selected ? 'bg-[#DDF4EB] border-[#147D6473]' : 'bg-[#ECF5F0] border-[#DDE8E0]'
                 }`}
               >
                 <AppText
@@ -145,7 +143,10 @@ export function ChatListHeader({
                 active ? 'bg-[#DDF4EB] border-[#147D6473]' : 'bg-[#ECF5F0] border-[#DDE8E0]'
               }`}
             >
-              <AppText variant="caption" className={`font-bold ${active ? 'text-[#147D64]' : 'text-[#5A7467]'}`}>
+              <AppText
+                variant="caption"
+                className={`font-bold ${active ? 'text-[#147D64]' : 'text-[#5A7467]'}`}
+              >
                 {label}
               </AppText>
             </ScalePressable>
@@ -161,7 +162,10 @@ export function ChatListHeader({
             </AppText>
             <AppText variant="caption" className="text-[#5A7467] font-bold">
               {activeFilterCount > 0
-                ? t('chat.active_filters', { count: activeFilterCount, defaultValue: '{{count}} filters' })
+                ? t('chat.active_filters', {
+                    count: activeFilterCount,
+                    defaultValue: '{{count}} filters',
+                  })
                 : agents.length}
             </AppText>
           </View>
@@ -173,7 +177,9 @@ export function ChatListHeader({
             <ScalePressable
               onPress={() => onSelectAgent(null)}
               className={`me-2 rounded-full px-3 py-2 border ${
-                selectedAgentId ? 'bg-[#ECF5F0] border-[#DDE8E0]' : 'bg-[#DDF4EB] border-[#147D6473]'
+                selectedAgentId
+                  ? 'bg-[#ECF5F0] border-[#DDE8E0]'
+                  : 'bg-[#DDF4EB] border-[#147D6473]'
               }`}
             >
               <AppText
