+++ frontend/src/components/productivity/workspace/index.ts
@@ -1,0 +1,6 @@
+export * from './WorkspaceHeader';
+export * from './WorkspaceTabs';
+export * from './TaskFilters';
+export * from './WorkspaceEmptyState';
+export * from './TodayPlanCard';
+export * from './EventCard';
