+++ frontend/src/components/settings/index.ts
@@ -1,0 +1,4 @@
+export * from './SectionHeader';
+export * from './SettingRow';
+export * from './MemoryItem';
+export * from './LanguagePickerModal';
