+++ frontend/src/core/i18n/nl.json
@@ -282,8 +282,7 @@
       "work": "Werk",
       "planning": "Planning",
       "creative": "Creatief"
-    }
-    ,
+    },
     "templatesComingSoon": "Templates komen binnenkort",
     "templatesComingSoonDesc": "Vooraf gebouwde persona's verschijnen hier.",
     "noTemplatesInCategory": "Geen templates in deze categorie",
